Organiza todos os metodos e classes em um diretorio para facilitar o acesso dos módulos
