+++ methods/classes/Instance.py
@@ -1,4 +1,3 @@
-import csv
 import math
 import operator
 
@@ -47,56 +46,3 @@ class Instance:
             dictClass[classification] += 1
 
         return max(dictClass.items(), key=operator.itemgetter(1))[0]
-
-# atualiza o array de minimo e maximo de cada um dos parametros
-def updateMinMax(row, minArg, maxArg):
-    row = [float(i) for i in row]
-    if (maxArg == []):
-        minArg = list(row)
-        maxArg = list(row)
-        return minArg, maxArg
-    else:
-        for i, param in enumerate(row):
-            if maxArg[i] < param:
-                maxArg[i] = param
-            if minArg[i] > param:
-                minArg[i] = param
-    return minArg, maxArg
-
-# ler do arquivo csv e salva as informacoes nos arrays
-def readCsv(file): 
-    classes = []
-    tests = []
-    maxArg = []
-    minArg = []
-
-    with open(file) as csvFile:
-        csvReader = csv.reader(csvFile, delimiter=',')
-        line_count = 0
-        for row in csvReader:
-            if (row != [] and line_count != 0):
-                param = [float(i) for i in row[:len(row)-1]]     
-                classification = row[len(row)-1]
-                test = Instance(line_count-1, param, classification)
-                tests.append(test)
-                minArg, maxArg = updateMinMax(row[:len(row)-1], minArg, maxArg)
-                if not(classification in classes):
-                    classes.append(classification)
-            line_count += 1
-    
-    return (classes, tests, minArg, maxArg)
-                
-# Retorna a classificação de uma instancia dado um nome de dataset e um k (número de vizinhos)
-def knn(classes, tests, minArg, maxArg, instance, k): 
-    
-    for inst in tests:
-        distanceToInstance = instance.euclideanDistance(inst, minArg, maxArg)
-        instance.insertDistance(distanceToInstance, inst, k)
-        
-    return instance.classify(k, classes)
-
-
-
-
-
-
